Removes lint warnings: unnecessary boxing, unboxing

+++ collect_app/src/main/java/org/odk/collect/android/widgets/ExDecimalWidget.java
@@ -50,7 +50,7 @@ public class ExDecimalWidget extends ExStringWidget {
             Object dataValue = dataHolder.getValue();
             if (dataValue != null) {
                 if (dataValue instanceof Integer) {
-                    d = Double.valueOf(((Integer) dataValue).intValue());
+                    d = (double) (Integer) dataValue;
                 } else {
                     d = (Double) dataValue;
                 }
